% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_rejections2}
\alias{add_rejections2}
\title{Annotate significance in one differential-abundance result}
\usage{
add_rejections2(diff, alpha = 0.05, lfc = 1, sigCol = "p.adj")
}
\arguments{
\item{diff}{A differential-analysis \code{SummarizedExperiment} object.}

\item{alpha}{A numeric p-value significance threshold.}

\item{lfc}{A numeric absolute log2 fold-change threshold.}

\item{sigCol}{One of \code{"p.val"}, \code{"p.adj"}, or \code{"BHCorrection"}, identifying
the significance column family.}
}
\value{
\code{diff} with contrast-specific significance columns added to
\code{rowData}.
}
\description{
Annotate significance in one differential-abundance result
}
\seealso{
Other proteomics functions:
\code{\link[=DEAnalysis]{DEAnalysis()}},
\code{\link[=DESEProtGenerate]{DESEProtGenerate()}},
\code{\link[=DataFilter]{DataFilter()}},
\code{\link[=DataImpute]{DataImpute()}},
\code{\link[=DesignMatrixFromNames]{DesignMatrixFromNames()}},
\code{\link[=DetermineMissing]{DetermineMissing()}},
\code{\link[=FormatMaxQuant]{FormatMaxQuant()}},
\code{\link[=LowSampleCountRemove]{LowSampleCountRemove()}},
\code{\link[=MultiNormalization]{MultiNormalization()}},
\code{\link[=NPeptideThreshold]{NPeptideThreshold()}},
\code{\link[=ProtSELoad]{ProtSELoad()}},
\code{\link[=ProteomicProteinName]{ProteomicProteinName()}},
\code{\link[=ProteomicSELoad]{ProteomicSELoad()}},
\code{\link[=ProteomicsDataDownload]{ProteomicsDataDownload()}},
\code{\link[=RowDataCompile]{RowDataCompile()}},
\code{\link[=SaveToProteomicDB]{SaveToProteomicDB()}},
\code{\link[=SigDEAnnotate]{SigDEAnnotate()}},
\code{\link[=impute2]{impute2()}},
\code{\link[=proteomicMZSave]{proteomicMZSave()}}
}
\concept{proteomics functions}
