% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DesignMatrixFromNames}
\alias{DesignMatrixFromNames}
\title{Build a proteomics design matrix from file names}
\usage{
DesignMatrixFromNames(Fpath)
}
\arguments{
\item{Fpath}{A path containing formatted proteomics files.}
}
\value{
A data table describing the experimental design.
}
\description{
Derives dataset, sample label, condition, and replicate fields from the file
names in a directory of formatted proteomics tables.
}
\seealso{
Other proteomics functions:
\code{\link[=DEAnalysis]{DEAnalysis()}},
\code{\link[=DESEProtGenerate]{DESEProtGenerate()}},
\code{\link[=DataFilter]{DataFilter()}},
\code{\link[=DataImpute]{DataImpute()}},
\code{\link[=DetermineMissing]{DetermineMissing()}},
\code{\link[=FormatMaxQuant]{FormatMaxQuant()}},
\code{\link[=LowSampleCountRemove]{LowSampleCountRemove()}},
\code{\link[=MultiNormalization]{MultiNormalization()}},
\code{\link[=NPeptideThreshold]{NPeptideThreshold()}},
\code{\link[=ProtSELoad]{ProtSELoad()}},
\code{\link[=ProteomicProteinName]{ProteomicProteinName()}},
\code{\link[=ProteomicSELoad]{ProteomicSELoad()}},
\code{\link[=ProteomicsDataDownload]{ProteomicsDataDownload()}},
\code{\link[=RowDataCompile]{RowDataCompile()}},
\code{\link[=SaveToProteomicDB]{SaveToProteomicDB()}},
\code{\link[=SigDEAnnotate]{SigDEAnnotate()}},
\code{\link[=add_rejections2]{add_rejections2()}},
\code{\link[=impute2]{impute2()}},
\code{\link[=proteomicMZSave]{proteomicMZSave()}}
}
\concept{proteomics functions}
