% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MultiNormalization}
\alias{MultiNormalization}
\title{Apply multiple proteomics normalizations}
\usage{
MultiNormalization(dataPepCutOff)
}
\arguments{
\item{dataPepCutOff}{A list of peptide-filtered \code{SummarizedExperiment}
objects.}
}
\value{
A named list whose elements (\code{mean}, \code{median}, \code{vsn}, \code{DEPvsn},
\code{loess}, \code{rlr}, and \code{smad}) each contain a normalized experiment list.
}
\description{
Applies mean, median, variance-stabilizing, cyclic loess, robust linear
regression, and robust mean/scale normalizations to facilitate method
comparison.
}
\seealso{
Other proteomics functions:
\code{\link[=DEAnalysis]{DEAnalysis()}},
\code{\link[=DESEProtGenerate]{DESEProtGenerate()}},
\code{\link[=DataFilter]{DataFilter()}},
\code{\link[=DataImpute]{DataImpute()}},
\code{\link[=DesignMatrixFromNames]{DesignMatrixFromNames()}},
\code{\link[=DetermineMissing]{DetermineMissing()}},
\code{\link[=FormatMaxQuant]{FormatMaxQuant()}},
\code{\link[=LowSampleCountRemove]{LowSampleCountRemove()}},
\code{\link[=NPeptideThreshold]{NPeptideThreshold()}},
\code{\link[=ProtSELoad]{ProtSELoad()}},
\code{\link[=ProteomicProteinName]{ProteomicProteinName()}},
\code{\link[=ProteomicSELoad]{ProteomicSELoad()}},
\code{\link[=ProteomicsDataDownload]{ProteomicsDataDownload()}},
\code{\link[=RowDataCompile]{RowDataCompile()}},
\code{\link[=SaveToProteomicDB]{SaveToProteomicDB()}},
\code{\link[=SigDEAnnotate]{SigDEAnnotate()}},
\code{\link[=add_rejections2]{add_rejections2()}},
\code{\link[=impute2]{impute2()}},
\code{\link[=proteomicMZSave]{proteomicMZSave()}}
}
\concept{proteomics functions}
