% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiOmicsDataCompile-package.R
\docType{package}
\name{MultiOmicsDataCompile-package}
\alias{MultiOmicsDataCompile}
\alias{MultiOmicsDataCompile-package}
\title{MultiOmicsDataCompile: compile and explore multi-omics data}
\description{
MultiOmicsDataCompile provides data-ingestion and harmonization workflows for
public
transcriptomic and proteomic datasets, utilities for constructing integrated
\code{SummarizedExperiment} objects, and lightweight exploratory visualizations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brendangongol/MultiOmicsDataCompile}
  \item Report bugs at \url{https://github.com/brendangongol/MultiOmicsDataCompile/issues}
}

}
\author{
\strong{Maintainer}: Brendan Gongol \email{8220574+brendangongol@users.noreply.github.com}

Authors:
\itemize{
  \item Brendan Gongol \email{8220574+brendangongol@users.noreply.github.com}
  \item Xueyuan Jiang
  \item Xiaowen Chen
  \item Yang Liu
}

}
\keyword{internal}
