% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{NPeptideThreshold}
\alias{NPeptideThreshold}
\title{Filter proteins by peptide count}
\usage{
NPeptideThreshold(dataFiltList, Npeptides = 2)
}
\arguments{
\item{dataFiltList}{A list of \code{SummarizedExperiment} objects with a
\code{Numberofpeptides} row-data column.}

\item{Npeptides}{The minimum number of peptides required for a protein.}
}
\value{
A list with \code{data}, the filtered experiments, and \code{remaining}, a data
table containing the proportion of proteins retained per dataset.
}
\description{
Filter proteins by peptide count
}
\seealso{
Other proteomics functions:
\code{\link[=DEAnalysis]{DEAnalysis()}},
\code{\link[=DESEProtGenerate]{DESEProtGenerate()}},
\code{\link[=DataFilter]{DataFilter()}},
\code{\link[=DataImpute]{DataImpute()}},
\code{\link[=DesignMatrixFromNames]{DesignMatrixFromNames()}},
\code{\link[=DetermineMissing]{DetermineMissing()}},
\code{\link[=FormatMaxQuant]{FormatMaxQuant()}},
\code{\link[=LowSampleCountRemove]{LowSampleCountRemove()}},
\code{\link[=MultiNormalization]{MultiNormalization()}},
\code{\link[=ProtSELoad]{ProtSELoad()}},
\code{\link[=ProteomicProteinName]{ProteomicProteinName()}},
\code{\link[=ProteomicSELoad]{ProteomicSELoad()}},
\code{\link[=ProteomicsDataDownload]{ProteomicsDataDownload()}},
\code{\link[=RowDataCompile]{RowDataCompile()}},
\code{\link[=SaveToProteomicDB]{SaveToProteomicDB()}},
\code{\link[=SigDEAnnotate]{SigDEAnnotate()}},
\code{\link[=add_rejections2]{add_rejections2()}},
\code{\link[=impute2]{impute2()}},
\code{\link[=proteomicMZSave]{proteomicMZSave()}}
}
\concept{proteomics functions}
