% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ProteomicProteinName}
\alias{ProteomicProteinName}
\title{List protein names in saved proteomics datasets}
\usage{
ProteomicProteinName(fPath)
}
\arguments{
\item{fPath}{A path containing proteomic \code{SummarizedExperiment} RDS files.}
}
\value{
A sorted character vector of unique protein names.
}
\description{
List protein names in saved proteomics datasets
}
\seealso{
Other proteomics functions:
\code{\link[=DEAnalysis]{DEAnalysis()}},
\code{\link[=DESEProtGenerate]{DESEProtGenerate()}},
\code{\link[=DataFilter]{DataFilter()}},
\code{\link[=DataImpute]{DataImpute()}},
\code{\link[=DesignMatrixFromNames]{DesignMatrixFromNames()}},
\code{\link[=DetermineMissing]{DetermineMissing()}},
\code{\link[=FormatMaxQuant]{FormatMaxQuant()}},
\code{\link[=LowSampleCountRemove]{LowSampleCountRemove()}},
\code{\link[=MultiNormalization]{MultiNormalization()}},
\code{\link[=NPeptideThreshold]{NPeptideThreshold()}},
\code{\link[=ProtSELoad]{ProtSELoad()}},
\code{\link[=ProteomicSELoad]{ProteomicSELoad()}},
\code{\link[=ProteomicsDataDownload]{ProteomicsDataDownload()}},
\code{\link[=RowDataCompile]{RowDataCompile()}},
\code{\link[=SaveToProteomicDB]{SaveToProteomicDB()}},
\code{\link[=SigDEAnnotate]{SigDEAnnotate()}},
\code{\link[=add_rejections2]{add_rejections2()}},
\code{\link[=impute2]{impute2()}},
\code{\link[=proteomicMZSave]{proteomicMZSave()}}
}
\concept{proteomics functions}
