% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ProteomicSELoad}
\alias{ProteomicSELoad}
\title{Load a proteomics experiment}
\usage{
ProteomicSELoad(Path, Fname = NULL)
}
\arguments{
\item{Path}{A path containing saved proteomics RDS files.}

\item{Fname}{The filename of the dataset to load. When omitted, the first RDS
file in \code{Path} is used.}
}
\value{
A one-element named list containing the selected
\code{SummarizedExperiment} object.
}
\description{
Load a proteomics experiment
}
\seealso{
Other proteomics functions:
\code{\link[=DEAnalysis]{DEAnalysis()}},
\code{\link[=DESEProtGenerate]{DESEProtGenerate()}},
\code{\link[=DataFilter]{DataFilter()}},
\code{\link[=DataImpute]{DataImpute()}},
\code{\link[=DesignMatrixFromNames]{DesignMatrixFromNames()}},
\code{\link[=DetermineMissing]{DetermineMissing()}},
\code{\link[=FormatMaxQuant]{FormatMaxQuant()}},
\code{\link[=LowSampleCountRemove]{LowSampleCountRemove()}},
\code{\link[=MultiNormalization]{MultiNormalization()}},
\code{\link[=NPeptideThreshold]{NPeptideThreshold()}},
\code{\link[=ProtSELoad]{ProtSELoad()}},
\code{\link[=ProteomicProteinName]{ProteomicProteinName()}},
\code{\link[=ProteomicsDataDownload]{ProteomicsDataDownload()}},
\code{\link[=RowDataCompile]{RowDataCompile()}},
\code{\link[=SaveToProteomicDB]{SaveToProteomicDB()}},
\code{\link[=SigDEAnnotate]{SigDEAnnotate()}},
\code{\link[=add_rejections2]{add_rejections2()}},
\code{\link[=impute2]{impute2()}},
\code{\link[=proteomicMZSave]{proteomicMZSave()}}
}
\concept{proteomics functions}
