% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RowDataCompile}
\alias{RowDataCompile}
\title{Reshape assay data to long format}
\usage{
RowDataCompile(tissueSplitList)
}
\arguments{
\item{tissueSplitList}{A list of \code{SummarizedExperiment} objects.}
}
\value{
A data table containing gene symbols, sample values, and parsed cell,
tissue, and diet labels.
}
\description{
Reshape assay data to long format
}
\seealso{
Other proteomics functions:
\code{\link[=DEAnalysis]{DEAnalysis()}},
\code{\link[=DESEProtGenerate]{DESEProtGenerate()}},
\code{\link[=DataFilter]{DataFilter()}},
\code{\link[=DataImpute]{DataImpute()}},
\code{\link[=DesignMatrixFromNames]{DesignMatrixFromNames()}},
\code{\link[=DetermineMissing]{DetermineMissing()}},
\code{\link[=FormatMaxQuant]{FormatMaxQuant()}},
\code{\link[=LowSampleCountRemove]{LowSampleCountRemove()}},
\code{\link[=MultiNormalization]{MultiNormalization()}},
\code{\link[=NPeptideThreshold]{NPeptideThreshold()}},
\code{\link[=ProtSELoad]{ProtSELoad()}},
\code{\link[=ProteomicProteinName]{ProteomicProteinName()}},
\code{\link[=ProteomicSELoad]{ProteomicSELoad()}},
\code{\link[=ProteomicsDataDownload]{ProteomicsDataDownload()}},
\code{\link[=SaveToProteomicDB]{SaveToProteomicDB()}},
\code{\link[=SigDEAnnotate]{SigDEAnnotate()}},
\code{\link[=add_rejections2]{add_rejections2()}},
\code{\link[=impute2]{impute2()}},
\code{\link[=proteomicMZSave]{proteomicMZSave()}}
}
\concept{proteomics functions}
