% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SigDEAnnotate}
\alias{SigDEAnnotate}
\title{Annotate significant differential abundance}
\usage{
SigDEAnnotate(dataDiff, alpha = 0.05, lfc = log2(2), sigCol = "p.adj")
}
\arguments{
\item{dataDiff}{A named list of differential-analysis
\code{SummarizedExperiment} objects.}

\item{alpha}{A numeric p-value significance threshold.}

\item{lfc}{A numeric absolute log2 fold-change threshold.}

\item{sigCol}{One of \code{"p.val"}, \code{"p.adj"}, or \code{"BHCorrection"}, identifying
the significance column family.}
}
\value{
A named list of experiments with significance columns added to
\code{rowData}.
}
\description{
Annotate significant differential abundance
}
\seealso{
Other proteomics functions:
\code{\link[=DEAnalysis]{DEAnalysis()}},
\code{\link[=DESEProtGenerate]{DESEProtGenerate()}},
\code{\link[=DataFilter]{DataFilter()}},
\code{\link[=DataImpute]{DataImpute()}},
\code{\link[=DesignMatrixFromNames]{DesignMatrixFromNames()}},
\code{\link[=DetermineMissing]{DetermineMissing()}},
\code{\link[=FormatMaxQuant]{FormatMaxQuant()}},
\code{\link[=LowSampleCountRemove]{LowSampleCountRemove()}},
\code{\link[=MultiNormalization]{MultiNormalization()}},
\code{\link[=NPeptideThreshold]{NPeptideThreshold()}},
\code{\link[=ProtSELoad]{ProtSELoad()}},
\code{\link[=ProteomicProteinName]{ProteomicProteinName()}},
\code{\link[=ProteomicSELoad]{ProteomicSELoad()}},
\code{\link[=ProteomicsDataDownload]{ProteomicsDataDownload()}},
\code{\link[=RowDataCompile]{RowDataCompile()}},
\code{\link[=SaveToProteomicDB]{SaveToProteomicDB()}},
\code{\link[=add_rejections2]{add_rejections2()}},
\code{\link[=impute2]{impute2()}},
\code{\link[=proteomicMZSave]{proteomicMZSave()}}
}
\concept{proteomics functions}
