% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{densityPlotFromList}
\alias{densityPlotFromList}
\title{Plot assay-value densities}
\usage{
densityPlotFromList(MultiNormalizeList)
}
\arguments{
\item{MultiNormalizeList}{A named list of normalization results, where each
element is a list of \code{SummarizedExperiment} objects.}
}
\value{
A named list of \code{ggplot} density plots.
}
\description{
Plot assay-value densities
}
\seealso{
Other exploration functions:
\code{\link[=VolcanoPlot]{VolcanoPlot()}},
\code{\link[=plot_volcano2]{plot_volcano2()}}
}
\concept{exploration functions}
