% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LowSampleCountRemove}
\alias{LowSampleCountRemove}
\alias{LowSampleCountRmove}
\title{Remove datasets with too few quantified features}
\usage{
LowSampleCountRemove(VarRMList, cut)

LowSampleCountRmove(VarRMList, cut)
}
\arguments{
\item{VarRMList}{A named list of \code{SummarizedExperiment} objects.}

\item{cut}{The minimum number of assay rows required to retain a dataset.}
}
\value{
The subset of \code{VarRMList} whose experiments contain more than \code{cut}
features.
}
\description{
Remove datasets with too few quantified features
}
\seealso{
Other proteomics functions:
\code{\link[=DEAnalysis]{DEAnalysis()}},
\code{\link[=DESEProtGenerate]{DESEProtGenerate()}},
\code{\link[=DataFilter]{DataFilter()}},
\code{\link[=DataImpute]{DataImpute()}},
\code{\link[=DesignMatrixFromNames]{DesignMatrixFromNames()}},
\code{\link[=DetermineMissing]{DetermineMissing()}},
\code{\link[=FormatMaxQuant]{FormatMaxQuant()}},
\code{\link[=MultiNormalization]{MultiNormalization()}},
\code{\link[=NPeptideThreshold]{NPeptideThreshold()}},
\code{\link[=ProtSELoad]{ProtSELoad()}},
\code{\link[=ProteomicProteinName]{ProteomicProteinName()}},
\code{\link[=ProteomicSELoad]{ProteomicSELoad()}},
\code{\link[=ProteomicsDataDownload]{ProteomicsDataDownload()}},
\code{\link[=RowDataCompile]{RowDataCompile()}},
\code{\link[=SaveToProteomicDB]{SaveToProteomicDB()}},
\code{\link[=SigDEAnnotate]{SigDEAnnotate()}},
\code{\link[=add_rejections2]{add_rejections2()}},
\code{\link[=impute2]{impute2()}},
\code{\link[=proteomicMZSave]{proteomicMZSave()}}
}
\concept{proteomics functions}
