% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_volcano2}
\alias{plot_volcano2}
\title{Create a differential-abundance volcano plot}
\usage{
plot_volcano2(
  dep,
  contrast,
  label_size = 3,
  add_names = TRUE,
  adjusted = FALSE,
  plot = TRUE,
  BHadjusted = FALSE
)
}
\arguments{
\item{dep}{A differential-analysis \code{SummarizedExperiment} object.}

\item{contrast}{A single comparison name present in the object's row data.}

\item{label_size}{The numeric size of feature labels.}

\item{add_names}{A logical value indicating whether significant feature names
should be added.}

\item{adjusted}{A logical value indicating whether DEP-adjusted p-values
should be used.}

\item{plot}{A logical value indicating whether to return a plot (\code{TRUE}) or
the underlying plotting data (\code{FALSE}).}

\item{BHadjusted}{A logical value indicating whether separately calculated
Benjamini-Hochberg-adjusted p-values should be used.}
}
\value{
A \code{ggplot} object when \code{plot = TRUE}; otherwise, a data frame with
protein, fold-change, p-value, and significance columns.
}
\description{
Create a differential-abundance volcano plot
}
\seealso{
Other exploration functions:
\code{\link[=VolcanoPlot]{VolcanoPlot()}},
\code{\link[=densityPlotFromList]{densityPlotFromList()}}
}
\concept{exploration functions}
